src folder renamed to source (Xamarin CI structure convention)
