--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "Xamarin.Auth"]
-	path = Xamarin.Auth
-	url = https://github.com/xamarin/Xamarin.Auth.git
