+++ source/Xamarin.Social.Android/ToolbarView.cs
