+++ source/Xamarin.Social.iOS/CheckedPickerView.cs
