+++ source/Xamarin.Social.iOS/NSDataStream.cs
