+++ source/Xamarin.Social.iOS/ProgressLabel.cs
