+++ source/Xamarin.Social/AssemblyInfo.cs
