+++ source/Xamarin.Social/FileData.cs
