+++ source/Xamarin.Social/ImageData.cs
