+++ source/Xamarin.Social/Item.cs
