+++ source/Xamarin.Social/Services/AppDotNetService.cs
