+++ source/Xamarin.Social/ShareResult.cs
