+++ source/Xamarin.Social/SocialException.cs
