+++ source/Xamarin.Social/Xamarin.Social.csproj
