--- src/Xamarin.Social.iOS/Xamarin.Social.iOS-Classic.csproj
@@ -1,121 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProductVersion>8.0.30703</ProductVersion>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{C91EF8FC-37E3-4952-A8A3-BA3A4422B0C5}</ProjectGuid>
-    <ProjectTypeGuids>{6BC8ED88-2882-458C-8E55-DFD12B67127B};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <OutputType>Library</OutputType>
-    <RootNamespace>Xamarin.Social</RootNamespace>
-    <AssemblyName>Xamarin.Social.iOS</AssemblyName>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>True</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>False</Optimize>
-    <DefineConstants>DEBUG;PLATFORM_IOS</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>False</ConsolePause>
-    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
-    <GenerateDocumentation>True</GenerateDocumentation>
-    <IntermediateOutputPath>obj\classic\AnyCPU\Debug\</IntermediateOutputPath>
-    <OutputPath>bin\classic\AnyCPU\Debug\</OutputPath>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>none</DebugType>
-    <Optimize>True</Optimize>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>False</ConsolePause>
-    <DefineConstants>PLATFORM_IOS</DefineConstants>
-    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
-    <GenerateDocumentation>True</GenerateDocumentation>
-    <IntermediateOutputPath>obj\classic\AnyCPU\Release\</IntermediateOutputPath>
-    <OutputPath>bin\classic\AnyCPU\Release\</OutputPath>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Xml" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Json" />
-    <Reference Include="monotouch" />
-    <Reference Include="Xamarin.Mobile">
-      <HintPath>..\Components\xamarin.mobile-0.7.5\lib\ios\Xamarin.Mobile.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-  <Import Project="$(MSBuildExtensionsPath)\Xamarin\iOS\Xamarin.MonoTouch.CSharp.targets" />
-  <ItemGroup>
-    <Compile Include="..\Xamarin.Social\AssemblyInfo.cs">
-      <Link>AssemblyInfo.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\Item.cs">
-      <Link>Item.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\Service.cs">
-      <Link>Service.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\ShareResult.cs">
-      <Link>ShareResult.cs</Link>
-    </Compile>
-    <Compile Include="FoundationResponse.cs" />
-    <Compile Include="..\Xamarin.Social\Services\FacebookService.cs">
-      <Link>Services\FacebookService.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\Services\TwitterService.cs">
-      <Link>Services\TwitterService.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\ImageData.cs">
-      <Link>ImageData.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\FileData.cs">
-      <Link>FileData.cs</Link>
-    </Compile>
-    <Compile Include="NSDataStream.cs" />
-    <Compile Include="CheckedPickerView.cs" />
-    <Compile Include="..\Xamarin.Social\Services\FlickrService.cs">
-      <Link>Services\FlickrService.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\OAuth1Service.cs">
-      <Link>OAuth1Service.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\SocialException.cs">
-      <Link>SocialException.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Social\OAuth2Service.cs">
-      <Link>OAuth2Service.cs</Link>
-    </Compile>
-    <Compile Include="ProgressLabel.cs" />
-    <Compile Include="ShareViewController.cs" />
-    <Compile Include="..\Xamarin.Social\Services\AppDotNetService.cs">
-      <Link>Services\AppDotNetService.cs</Link>
-    </Compile>
-    <Compile Include="Services\Twitter5Service.cs" />
-    <Compile Include="..\..\Xamarin.Auth\src\Xamarin.Auth.iOS\UIViewControllerEx.cs">
-      <Link>UIViewControllerEx.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Xamarin.Auth\src\Xamarin.Auth\ExceptionEx.cs">
-      <Link>ExceptionEx.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Xamarin.Auth\src\Xamarin.Auth\WebEx.cs">
-      <Link>WebEx.cs</Link>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Services\" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\Xamarin.Auth\src\Xamarin.Auth.iOS\Xamarin.Auth.iOS-Classic.csproj">
-      <Project>{9309BCCC-AA7B-4195-A130-D4572FA9488D}</Project>
-      <Name>Xamarin.Auth.iOS-Classic</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <XamarinComponentReference Include="xamarin.mobile">
-      <Version>0.7.5</Version>
-      <Visible>False</Visible>
-    </XamarinComponentReference>
-  </ItemGroup>
-</Project>
