--- src/Xamarin.Social/Xamarin.Social.sln
@@ -1,26 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xamarin.Social", "Xamarin.Social.csproj", "{2CAFD7AE-C52F-4F01-9E25-58656DD8476D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xamarin.Auth", "..\..\Xamarin.Auth\src\Xamarin.Auth\Xamarin.Auth.csproj", "{C9FF1919-78DD-4084-976D-BDABC0A7E80B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2CAFD7AE-C52F-4F01-9E25-58656DD8476D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2CAFD7AE-C52F-4F01-9E25-58656DD8476D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2CAFD7AE-C52F-4F01-9E25-58656DD8476D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2CAFD7AE-C52F-4F01-9E25-58656DD8476D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C9FF1919-78DD-4084-976D-BDABC0A7E80B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C9FF1919-78DD-4084-976D-BDABC0A7E80B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C9FF1919-78DD-4084-976D-BDABC0A7E80B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C9FF1919-78DD-4084-976D-BDABC0A7E80B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = Xamarin.Social.csproj
-	EndGlobalSection
-EndGlobal
